+++ packages/@n8n/backend-common/src/license-state.ts
@@ -26,9 +26,12 @@ export class LicenseState {
 	// --------------------
 	//     core queries
 	// --------------------
-
-	isLicensed(feature: BooleanLicenseFeature) {
+	isLicensed(feature: BooleanLicenseFeature | BooleanLicenseFeature[]) {
 		this.assertProvider();
+
+		if (Array.isArray(feature)) {
+			return feature.some(this.licenseProvider.isLicensed);
+		}
 
 		return this.licenseProvider.isLicensed(feature);
 	}
