+++ packages/cli/src/modules/provisioning.ee/__tests__/provisioning.controller.ee.test.ts
@@ -2,7 +2,7 @@ import type { LicenseState } from '@n8n/backend-common';
 import { mock } from 'jest-mock-extended';
 
 import { ProvisioningController } from '../provisioning.controller.ee';
-import { type ProvisioningService } from '@/provisioning.ee/provisioning.service.ee';
+import { type ProvisioningService } from '@/modules/provisioning.ee/provisioning.service.ee';
 import { type Response } from 'express';
 import { type AuthenticatedRequest } from '@n8n/db';
 import { type ProvisioningConfigDto } from '@n8n/api-types';
