+++ packages/cli/src/modules/provisioning.ee/__tests__/provisioning.service.ee.test.ts
@@ -1,7 +1,7 @@
 import type { Logger } from '@n8n/backend-common';
 import { mock } from 'jest-mock-extended';
 
-import { ProvisioningService } from '@/provisioning.ee/provisioning.service.ee';
+import { ProvisioningService } from '@/modules/provisioning.ee/provisioning.service.ee';
 import { type SettingsRepository } from '@n8n/db';
 import { type GlobalConfig } from '@n8n/config';
 import { PROVISIONING_PREFERENCES_DB_KEY } from '../constants';
