+++ packages/cli/src/modules/provisioning.ee/provisioning.module.ts
@@ -1,0 +1,9 @@
+import type { ModuleInterface } from '@n8n/decorators';
+import { BackendModule } from '@n8n/decorators';
+
+@BackendModule({ name: 'provisioning', licenseFlag: ['feat:oidc', 'feat:saml', 'feat:ldap'] })
+export class ExternalSecretsModule implements ModuleInterface {
+	async init() {
+		await import('./provisioning.controller.ee');
+	}
+}
